pyproject.toml
[project]
name = "scrubid"
version = "0.1.0"
description = "ScrubID implementation (ICLR pack v1.0.3)"
requires-python = ">=3.12"
dependencies = [
  "pyyaml>=6.0.1",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "scipy>=1.11.0",
  "matplotlib>=3.8.0",
  "tqdm>=4.66.0",
  "torch>=2.2.0",
  "transformers>=4.41.0",
  "transformer-lens>=2.6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[build-system]
requires = ["setuptools>=69.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

